feat: Enhance user registration with email and points; refactor routing structure
