+++ controllers/authController.js
@@ -1,13 +1,19 @@
-import bcrypt from "bcrypt";
-import jwt from "jsonwebtoken";
-import User from "../models/user";
+const bcrypt = require("bcrypt");
+const jwt = require("jsonwebtoken");
+const User = require("../models/user");
 
 const authController = {
   register: async (req, res) => {
     try {
-      const { username, password } = req.body;
+      const { username, password, email, points } = req.body;
       const hashedPassword = await bcrypt.hash(password, 10);
-      const user = new User({ username, password: hashedPassword });
+      const user = new User({
+        username,
+        password: hashedPassword,
+        email,
+        points,
+        createdAt: new Date(),
+      });
       await user.save();
       res.status(201).json({ message: "User registered successfully" });
     } catch (error) {
