+++ routes/auth.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const controller = require("../controllers/appController");
+const controller = require("../controllers/authController");
 
 router.post("/register", controller.register);
 router.post("/login", controller.login);
