+++ routes/index.js
@@ -1,0 +1,7 @@
+const authRoutes = require("./auth");
+const appRoutes = require("./app");
+
+module.exports = {
+  authRoutes,
+  appRoutes,
+};
